+++ metadata.rb
@@ -1,9 +1,11 @@
-name 'skeleton'
+name 'bamboo-agent'
 maintainer 'Numergy'
 maintainer_email 'cd@numergy.com'
 license 'Numergy'
-description 'Installs/Configures skeleton'
+description 'Installs/Configures a bamboo agent'
 long_description IO.read(File.join(File.dirname(__FILE__), 'README.md'))
 version '0.1.0'
 
 depends 'apt'
+depends 'augeas'
+depends 'java'