+++ spec/spec_helper.rb
@@ -5,6 +5,6 @@
 require 'chefspec'
 require 'chefspec/berkshelf'
 
-ChefSpec::Coverage.start! { add_filter 'skeleton' }
+ChefSpec::Coverage.start! { add_filter 'bamboo-agent' }
 
 require 'chef/application'