+++ CHANGELOG.md
@@ -1,11 +1,11 @@
-skeleton CHANGELOG
+Bamboo-agent CHANGELOG
 =================
 
-This file is used to list changes made in each version of the skeleton cookbook.
+This file is used to list changes made in each version of the bamboo-agent cookbook.
 
 0.1.0
 -----
-- [your_name] - Initial release of skeleton
+- [your_name] - Initial release of bamboo-agent
 
 - - -
 Check the [Markdown Syntax Guide](http://daringfireball.net/projects/markdown/syntax) for help with Markdown.